Efficient cross-attention mixing by allowing attention views (#85)
This commit adds attention stride arguments to every kernel, forcing
them to allow non-contiguous / view tensors for attention weights.

The goal is to be able to skip copying when concatenating neighborhood
attention weights with cross-attention weights. If there are N
neighborhood attention weights, and M cross-attention weights, we can
concatenate them without copying if we force the output of the
neighborhood attention and cross attention ops to be a strided tensor.
Cross attention is just a BMM, and torch.bmm allows strided outputs. NA
did not, but this commit allows it by setting attention strides to the
stride values from the torch tensor, and not based on problem size.

This commit also moves finding the default cuda stream to the torch API,
and passes the stream down correctly to both naive and GEMM kernels.

+++ csrc/include/natten/cuda/naive/tiled/pointwise_neighborhood_2d_tiled_11x11_13x13.cuh
@@ -89,7 +89,7 @@ struct PointwiseNeighborhood2DFull11x11_13x13
     // if (z < batch_size * heads)
     // {
     const int lti = threadIdx.y * (TILEY * KERNEL_SIZE) + threadIdx.x;
-    const int batchHeadOffset = b * p.query_stride_0 + h * p.query_stride_1;
+    const int64_t batchHeadOffset = b * p.query_stride_0 + h * p.query_stride_1;
     const int si = int(blockIdx.y / dilation) * (TILEX * dilation) +
         (blockIdx.y % dilation);
     const int sj = int(blockIdx.x / dilation) * (TILEY * dilation) +
@@ -121,7 +121,7 @@ struct PointwiseNeighborhood2DFull11x11_13x13
       const int bj = (ktx - bi * KTILEY) * dilation + snj;
       bi = bi * dilation + sni;
       if (bi < p.height && bj < p.width) {
-        const int keyOffset = batchHeadOffset + bi * p.query_stride_2 +
+        const int64_t keyOffset = batchHeadOffset + bi * p.query_stride_2 +
             bj * p.query_stride_3 + kty;
 #pragma unroll
         for (int ti = 0; ti < KITERS_32; ++ti)
@@ -148,14 +148,14 @@ struct PointwiseNeighborhood2DFull11x11_13x13
       for (int dimOffset = 0; dimOffset < DIM_32; ++dimOffset)
         updt += tile[queryIdx][dimOffset] * kTile[keyIdx][dimOffset];
 
-      const int index = b * p.attn_stride_0 + h * p.attn_stride_1 +
+      const int64_t index = b * p.attn_stride_0 + h * p.attn_stride_1 +
           i * p.attn_stride_2 + j * p.attn_stride_3 + ki * KERNEL_SIZE + kj;
       if (p.bias) {
         const int pi =
             get_pb_start(i, p.height, KERNEL_SIZE, NEIGHBORHOOD_SIZE, dilation);
         const int pj =
             get_pb_start(j, p.width, KERNEL_SIZE, NEIGHBORHOOD_SIZE, dilation);
-        const int biasIndex =
+        const int64_t biasIndex =
             h * p.bias_stride_0 + (pi + ki) * p.bias_stride_1 + (pj + kj);
         updt += p.bias[biasIndex];
       }
@@ -224,7 +224,7 @@ struct PointwiseNeighborhood2DHalf11x11_13x13
     // {
     const int lti = threadIdx.y * (TILEY * KERNEL_SIZE) + threadIdx.x;
     const int stride2 = DIMHALF_32 * p.width;
-    const int batchHeadOffset = b * p.query_stride_0 + h * p.query_stride_1;
+    const int64_t batchHeadOffset = b * p.query_stride_0 + h * p.query_stride_1;
     const int si = int(blockIdx.y / dilation) * (TILEX * dilation) +
         (blockIdx.y % dilation);
     const int sj = int(blockIdx.x / dilation) * (TILEY * dilation) +
@@ -260,7 +260,7 @@ struct PointwiseNeighborhood2DHalf11x11_13x13
       const int bj = (ktx - bi * KTILEY) * dilation + snj;
       bi = bi * dilation + sni;
       if (bi < p.height && bj < p.width) {
-        const int keyOffset =
+        const int64_t keyOffset =
             batchHeadOffset + bi * stride2 + bj * DIMHALF_32 + kty;
 #pragma unroll
         for (int ti = 0; ti < KHALFITERS_32; ++ti)
@@ -292,7 +292,7 @@ struct PointwiseNeighborhood2DHalf11x11_13x13
               kTile[keyIdx][di * KITERS_32 + dj],
               updt);
 
-      const int index = b * p.attn_stride_0 + h * p.attn_stride_1 +
+      const int64_t index = b * p.attn_stride_0 + h * p.attn_stride_1 +
           i * p.attn_stride_2 + j * p.attn_stride_3 + ki * KERNEL_SIZE + kj;
       scalar_t acc = HalfHelper::cast_back(HalfHelper::add(updt.x, updt.y));
       if (p.bias) {
@@ -300,7 +300,7 @@ struct PointwiseNeighborhood2DHalf11x11_13x13
             get_pb_start(i, p.height, KERNEL_SIZE, NEIGHBORHOOD_SIZE, dilation);
         const int pj =
             get_pb_start(j, p.width, KERNEL_SIZE, NEIGHBORHOOD_SIZE, dilation);
-        const int biasIndex =
+        const int64_t biasIndex =
             h * p.bias_stride_0 + (pi + ki) * p.bias_stride_1 + (pj + kj);
         acc = HalfHelper::add(acc, p.bias[biasIndex]);
       }
